refactor: LottoSeller 패키지 변경
- 의존성의 흐름을 관리 하기 위해 변경
