+++ src/main/java/lotto/factory/LottoSeller.java
@@ -1,5 +1,8 @@
-package lotto.domain;
+package lotto.factory;
 
+import lotto.domain.Lotto;
+import lotto.domain.LottoNumber;
+import lotto.domain.Lottos;
 import lotto.domain.strategy.ManualLottoNumberStrategy;
 import lotto.domain.strategy.RandomLottoNumberStrategy;
