+++ src/test/java/lotto/domain/LottoSellerTest.java
@@ -1,6 +1,7 @@
 package lotto.domain;
 
 import lotto.domain.strategy.RandomLottoNumberStrategy;
+import lotto.factory.LottoSeller;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
